Multiple issues to keep command simple:
to be able to call the utility system wide (avoiding having to copy
the script in each and every location you would use it from) I had
to find the root of the git project that the input was set from.

Finding the .git folder (reversing backwards recursively) is beyond
my abilities, so I found a lib called gitpath that does exactly
this for me.

Gitpath (https://github.com/maxnoe/python-gitpath) is written in
Python3 however so I had to undergo a bunch of translations I had
no idea about, and install Pip3 etc etc.

Before all that I did go back to GitPython a bit, it has a utility
to find git roots, but it's only part of an index function that
depends on the main Repo object, that needs to be set manually.
I asked about this on SO:
http://stackoverflow.com/questions/39645916/finding-git-root-from-file-with-gitpython

Python3 conversion done anyway, and now it all works, I think.